agent/requirements.txt
livekit-agents[openai,silero,deepgram]>=0.9.0
python-dotenv>=1.0.0
httpx>=0.25.0
